\name{plotlinc-methods}
\docType{methods}
\alias{plotlinc}
\alias{plotlinc-methods}
\alias{plotlinc,LINCbio,character-method}
\alias{plotlinc,LINCbio,missing-method}
\alias{plotlinc,LINCcluster,character-method}
\alias{plotlinc,LINCcluster,missing-method}
\alias{plotlinc,LINCmatrix,character-method}
\alias{plotlinc,LINCmatrix,missing-method}
\alias{plotlinc,LINCsingle,character-method}
\alias{plotlinc,LINCsingle,missing-method}
\title{Plot Objects Of The 'LINC' class}
\description{
\code{plotlinc} is the generic function that creates plots based on objects of the \pkg{LINC} class.
}
\section{Methods}{
\describe{

\item{\code{signature(input = "LINCbio", showCor = "character")}}{
(see details)
}

\item{\code{signature(input = "LINCbio", showCor = "missing")}}{
(see details)
}

\item{\code{signature(input = "LINCcluster", showCor = "character")}}{
(see details)
}

\item{\code{signature(input = "LINCcluster", showCor = "missing")}}{
(see details)
}

\item{\code{signature(input = "LINCmatrix", showCor = "character")}}{
(see details)
}

\item{\code{signature(input = "LINCmatrix", showCor = "missing")}}{
(see details)
}

\item{\code{signature(input = "LINCsingle", showCor = "character")}}{
(see details)
}

\item{\code{signature(input = "LINCsingle", showCor = "missing")}}{
(see details)
}
}}

\usage{
plotlinc(input,
         showCor,
         returnDat = FALSE)
}

\arguments{
  \item{input}{
 usually a \code{LINCmatrix}, \code{LINCcluster}, \code{LINCbio} or \code{LINCsingle} object  
}
  \item{showCor}{
a \code{vector} of gene names; length 2 up to 6 elements; the first entry is the subject (see example) 
}
  \item{returnDat}{
whether to return the data used to create the plot \code{TRUE} or not \code{FALSE}
}
}
\details{
A plot will be created based on the information found in the \code{slots}: \code{history} and \code{linCenvir} of a \code{LINCmatrix}, \code{LINCcluster}, \code{LINCbio} or \code{LINCsingle} object.
Using the argument \code{showCor = ...} will output scatterplots for one query, the first entry in \code{...}, and up to five subjects (see example). This can be used to manually check the correlation of genes.
Individual or modified plots can be generated by returning the plotting data with \code{returnDat}
Unexpected extreme values may corrupt plots.
}

\value{
an object of the class 'gtable' containing multiple 'grobs' or a 'list' if \code{returnDat = TRUE}
}

\section{Compatibility}{
\code{plotlinc(LINCmatrix, ...)}, \code{plotlinc(LINCcluster, ...)},
\code{plotlinc(LINCbio, ...)},\code{plotlinc(LINCsingle, ...)}, ...  
}


\author{
Manuel Goepferich
}

\seealso{
\code{querycluster} ; 
\code{feature}
}
\examples{
data(BRAIN_EXPR)
plotlinc(crbl_matrix)
plotlinc(crbl_cluster)

# show correlations and expression; "647979" as query plus 4 subjects
plotlinc(crbl_cluster, showCor = c("647979", "6726", "3337", "3304" ,"3320"))
}  
\keyword{ ~plotlinc }



\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
